Added so that you  cannot delete yourself
Made it so that you can no longer delete yourself to prevent issue #104  from happening.

+++ app/Http/Controllers/Admin/ClientController.php
@@ -78,6 +78,9 @@ class ClientController extends Controller
 
     public function destroy(User $user)
     {
+        if ($user->id === auth()->user()->id) {
+            return redirect()->back()->with('error', 'You cannot delete yourself.');
+        }
         $user->delete();
 
         return redirect()->route('admin.clients');
